requirements.txt
langgraph 
langchain-openai 
qiskit 
qiskit-aer
qiskit[visualization]
matplotlib 
python-dotenv
